Arreglado conexion con telemetria, para BQ3.3
